Workaround for not being able to build in VS
Workaround for: #12653

This isn't a great fix for the this problem but it makes it possible to
build the solution. You'll still see errors when you first open the
solution but they will go away when VS does a restore.

We'll continue the discussion separately about the right fix for this.